requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@master
audioop-lts
jishaku @ git+https://github.com/scarletcafe/jishaku@master

asyncpg
googletrans